test/requirements.txt
pandas==2.1.4
matplotlib==3.8.2
Pygments==2.17.2
Jinja2==3.1.2
PyYAML==6.0.1
pdfminer.six==20231228
